% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{wemo_ON}
\alias{wemo_ON}
\title{Turn WeMo Switch ON}
\usage{
wemo_ON(ip)
}
\arguments{
\item{ip}{A characted string or vector of character strings representing the
IP address(es) of WeMo Switch smart plug(s).}
}
\value{
A \code{\link[tibble]{tibble}} with two columns:
  \describe{
    \item{ip}{The IP address of the WeMo Switch smart plug}
    \item{success}{A logical indicating whether the WeMo Switch was
                   successfully turned on. NA will be returned is no WeMo
                   Switch is detected at this IP.}
  }
}
\description{
Turn WeMo Switch smart plugs on.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
